requirements.txt
openai
python-dotenv
faiss-cpu
numpy
tiktoken
rank_bm25